Exported functions and types

Index

LowLevelParticleFilters.AdvancedParticleFilter

LowLevelParticleFilters.AuxiliaryParticleFilter

LowLevelParticleFilters.DAEUnscentedKalmanFilter

LowLevelParticleFilters.ExtendedKalmanFilter

LowLevelParticleFilters.KalmanFilter

LowLevelParticleFilters.ParticleFilter

LowLevelParticleFilters.TupleProduct

LowLevelParticleFilters.UnscentedKalmanFilter

LowLevelParticleFilters.commandplot

LowLevelParticleFilters.correct!

LowLevelParticleFilters.debugplot

LowLevelParticleFilters.densityplot

LowLevelParticleFilters.forward_trajectory

LowLevelParticleFilters.forward_trajectory

LowLevelParticleFilters.log_likelihood_fun

LowLevelParticleFilters.loglik

LowLevelParticleFilters.logsumexp!

LowLevelParticleFilters.mean_trajectory

LowLevelParticleFilters.metropolis

LowLevelParticleFilters.predict!

LowLevelParticleFilters.reset!

LowLevelParticleFilters.rk4

LowLevelParticleFilters.simulate

LowLevelParticleFilters.smooth

LowLevelParticleFilters.smooth

LowLevelParticleFilters.smoothed_cov

LowLevelParticleFilters.smoothed_mean

LowLevelParticleFilters.smoothed_trajs

LowLevelParticleFilters.update!

LowLevelParticleFilters.weigthed_cov

LowLevelParticleFilters.weigthed_mean

LowLevelParticleFilters.weigthed_mean

LowLevelParticleFilters.AdvancedParticleFilter — Method

AdvancedParticleFilter(Nparticles, dynamics, measurement, measurement_likelihood, dynamics_density, initial_density; p = SciMLBase.NullParameters(), kwargs...)

See the docs for more information: https://baggepinnen.github.io/LowLevelParticleFilters.jl/stable/#AdvancedParticleFilter-1

LowLevelParticleFilters.AuxiliaryParticleFilter — Method

AuxiliaryParticleFilter(args...; kwargs...)

Takes exactly the same arguments as ParticleFilter, or an instance of ParticleFilter.

LowLevelParticleFilters.DAEUnscentedKalmanFilter — Method

DAEUnscentedKalmanFilter(ukf; g, get_x_z, build_xz, xz0, threads=false)

An Unscented Kalman filter for differential-algebraic systems (DAE).

Ref: "Nonlinear State Estimation of Differential Algebraic Systems", Mandela, Rengaswamy, Narasimhan

Warning

This filter is still considered experimental and subject to change without respecting semantic versioning. Use at your own risk.

Arguments

ukf is a regular UnscentedKalmanFilter that contains dynamics(xz, u, p, t) that propagates the combined state xz(k) to xz(k+1) and a measurement function with signature (xz, u, p, t)

g(x, z, u, p, t) is a function that should fulfill g(x, z, u, p, t) = 0

get_x_z(xz) -> x, z is a function that decomposes xz into x and z

build_xz(x, z) is the inverse of get_x_z

xz0 the initial full state.

threads: If true, evaluates dynamics on sigma points in parallel. This typically requires the dynamics to be non-allocating (use StaticArrays) to improve performance.

Assumptions

The DAE dynamics is index 1 and can be written on the form

ẋ = f(x, z, u, p, t) # Differential equations
0 = g(x, z, u, p, t) # Algebraic equations
y = h(x, z, u, p, t) # Measurements

the measurements may be functions of both differential states x and algebraic variables z. Note, the actual dynamcis and measurement functions stored in the internal ukf should have signatures (xz, u, p, t), i.e., they take the combined state containing both x and z in a single vector as dictated by the function build_xz. It is only the function g that is assumed to actually have the signature g(x,z,u,p,t).

LowLevelParticleFilters.ExtendedKalmanFilter — Type

ExtendedKalmanFilter(kf, dynamics, measurement)

A nonlinear state estimator propagating uncertainty using linearization.

An extended Kalman filter takes a standard Kalman filter as well as dynamics and measurement functions. The filter will linearize the dynamics using ForwardDiff. The dynamics and measurement function are on the following form

x' = dynamics(x, u, p, t) + w
y  = measurement(x, u, p, t) + e

where w ~ N(0, R1), e ~ N(0, R2) and x(0) ~ d0

See also UnscentedKalmanFilter which is typically more accurate than ExtendedKalmanFilter. See KalmanFilter for detailed instructions on how to set up the Kalman filter kf.

LowLevelParticleFilters.KalmanFilter — Type

KalmanFilter(A,B,C,D,R1,R2,d0=MvNormal(R1); p = SciMLBase.NullParameters())

The matrices A,B,C,D define the dynamics

x' = Ax + Bu + w
y  = Cx + Du + e

where w ~ N(0, R1), e ~ N(0, R2) and x(0) ~ d0

The matrices can be time varying such that, e.g., A[:, :, t] contains the $A$ matrix at time index t. They can also be given as functions on the form

Afun(x, u, p, t) -> A

For maximum performance, provide statically sized matrices from StaticArrays.jl

LowLevelParticleFilters.ParticleFilter — Method

ParticleFilter(num_particles, dynamics, measurement, dynamics_density, measurement_density, initial_density; p = SciMLBase.NullParameters())

See the docs for more information: https://baggepinnen.github.io/LowLevelParticleFilters.jl/stable/#Particle-filter-1

LowLevelParticleFilters.TupleProduct — Type

TupleProduct(v::NTuple{N,UnivariateDistribution})

Create a product distribution where the individual distributions are stored in a tuple. Supports mixed/hybrid Continuous and Discrete distributions

LowLevelParticleFilters.UnscentedKalmanFilter — Type

UnscentedKalmanFilter(dynamics, measurement, R1, R2, d0=MvNormal(Matrix(R1)); p = SciMLBase.NullParameters(), ny, nu)

A nonlinear state estimator propagating uncertainty using the unscented transform.

The dynamics and measurement function are on the following form

x' = dynamics(x, u, p, t) + w
y  = measurement(x, u, p, t) + e

where w ~ N(0, R1), e ~ N(0, R2) and x(0) ~ d0

The matrices R1, R2 can be time varying such that, e.g., R1[:, :, t] contains the $R_1$ matrix at time index t. They can also be given as functions on the form

Rfun(x, u, p, t) -> R

For maximum performance, provide statically sized matrices from StaticArrays.jl

ny, nu indicate the number of outputs and inputs.

LowLevelParticleFilters.commandplot — Function

commandplot(pf, u, y, p=parameters(pf); kwargs...)

Produce a helpful plot. For customization options (kwargs...), see ?pplot. After each time step, a command from the user is requested.

q: quit

s n: step n steps

LowLevelParticleFilters.correct! — Function

ll, e = correct!(f, u, y, p = parameters(f), t = index(f))

Update state/covariance/weights based on measurement y, returns loglikelihood and prediction error (the error is always 0 for particle filters).

LowLevelParticleFilters.debugplot — Function

debugplot(pf, u, y, p=parameters(pf); runall=false, kwargs...)

Produce a helpful plot. For customization options (kwargs...), see ?pplot.

runall=false: if true, runs all time steps befor displaying (faster), if false, displays the plot after each time step.

The generated plot becomes quite heavy. Initially, try limiting your input to 100 time steps to verify that it doesn't crash.

LowLevelParticleFilters.densityplot — Function

densityplot(x,[w])

Plot (weighted) particles densities

LowLevelParticleFilters.forward_trajectory — Function

sol = forward_trajectory(kf::AbstractKalmanFilter, u::Vector, y::Vector, p=parameters(kf))

Run a Kalman filter forward

Returns a KalmanFilteringSolution: with the following

x: predictions

xt: filtered estimates

R: predicted covariance matrices

Rt: filter covariances

ll: loglik

sol can be plotted

plot(sol::KalmanFilteringSolution; plotx = true, plotxt=true, plotu=true, ploty=true)

LowLevelParticleFilters.forward_trajectory — Function

sol = forward_trajectory(pf, u::AbstractVector, y::AbstractVector, p=parameters(pf))

Run the particle filter for a sequence of inputs and measurements. Return a solution with x,w,we,ll = particles, weights, expweights and loglikelihood

If MonteCarloMeasurements.jl is loaded, you may transform the output particles to Matrix{MonteCarloMeasurements.Particles} using Particles(x,we). Internally, the particles are then resampled such that they all have unit weight. This is conventient for making use of the plotting facilities of MonteCarloMeasurements.jl.

sol can be plotted

plot(sol::ParticleFilteringSolution; nbinsy=30, xreal=nothing, dim=nothing)

LowLevelParticleFilters.log_likelihood_fun — Method

ll(θ) = log_likelihood_fun(filter_from_parameters(θ::Vector)::Function, priors::Vector{Distribution}, u, y, p)

returns function θ -> p(y|θ)p(θ)

LowLevelParticleFilters.loglik — Function

ll = loglik(filter, u, y, p=parameters(filter))

Calculate loglikelihood for entire sequences u,y

LowLevelParticleFilters.logsumexp! — Function

ll = logsumexp!(w, we [, maxw])

Normalizes the weight vector w and returns the weighted log-likelihood

https://arxiv.org/pdf/1412.8695.pdf eq 3.8 for p(y) https://discourse.julialang.org/t/fast-logsumexp/22827/7?u=baggepinnen for stable logsumexp

LowLevelParticleFilters.mean_trajectory — Method

x,ll = mean_trajectory(pf, u::Vector{Vector}, y::Vector{Vector}, p=parameters(pf))

This Function resets the particle filter to the initial state distribution upon start

LowLevelParticleFilters.metropolis — Function

metropolis(ll::Function(θ), R::Int, θ₀::Vector, draw::Function(θ) = naive_sampler(θ₀))

Performs MCMC sampling using the marginal Metropolis (-Hastings) algorithm draw = θ -> θ' samples a new parameter vector given an old parameter vector. The distribution must be symmetric, e.g., a Gaussian. R is the number of iterations. See log_likelihood_fun

Example:

filter_from_parameters(θ) = ParticleFilter(N, dynamics, measurement, MvNormal(n,exp(θ[1])), MvNormal(p,exp(θ[2])), d0)
priors = [Normal(0,0.1),Normal(0,0.1)]
ll     = log_likelihood_fun(filter_from_parameters,priors,u,y,1)
θ₀ = log.([1.,1.]) # Initial point
draw = θ -> θ .+ rand(MvNormal(0.1ones(2))) # Function that proposes new parameters (has to be symmetric)
burnin = 200 # If using threaded call, provide number of burnin iterations
# @time theta, lls = metropolis(ll, 2000, θ₀, draw) # Run single threaded
# thetam = reduce(hcat, theta)'
@time thetalls = LowLevelParticleFilters.metropolis_threaded(burnin, ll, 5000, θ₀, draw) # run on all threads, will provide (2000-burnin)*nthreads() samples
histogram(exp.(thetalls[:,1:2]), layout=3)
plot!(thetalls[:,3], subplot=3) # if threaded call, log likelihoods are in the last column

LowLevelParticleFilters.predict! — Function

predict!(f, u, p = parameters(f), t = index(f))

Move filter state forward in time using dynamics equation and input vector u.

LowLevelParticleFilters.reset! — Method

Reset the filter to initial state and covariance/distribution

LowLevelParticleFilters.simulate — Function

x,u,y = simulate(f::AbstractFilter, T::Int, du::Distribution, p=parameters(f), [N]; dynamics_noise=true)
x,u,y = simulate(f::AbstractFilter, u, p=parameters(f); dynamics_noise=true)

Simulate dynamical system forward in time T steps, or for the duration of u, returns state sequence, inputs and measurements du is a distribution of random inputs.

If MonteCarloMeasurements.jl is loaded, the argument N::Int can be supplied, in which case N simulations are done and the result is returned in the form of Vector{MonteCarloMeasurements.Particles}.

LowLevelParticleFilters.smooth — Function

xT,RT,ll = smooth(kf::KalmanFilter, u::Vector, y::Vector, p=parameters(kf))

Returns smoothed estimates of state x and covariance R given all input output data u,y

LowLevelParticleFilters.smooth — Function

xb,ll = smooth(pf, M, u, y, p=parameters(pf))
xb,ll = smooth(pf, xf, wf, wef, ll, M, u, y, p=parameters(pf))

Perform particle smoothing using forward-filtering, backward simulation. Return smoothed particles and loglikelihood. See also smoothed_trajs, smoothed_mean, smoothed_cov

LowLevelParticleFilters.smoothed_cov — Method

smoothed_cov(xb)

Helper function to calculate the covariance of smoothed particle trajectories

LowLevelParticleFilters.smoothed_mean — Method

smoothed_mean(xb)

Helper function to calculate the mean of smoothed particle trajectories

LowLevelParticleFilters.smoothed_trajs — Method

smoothed_trajs(xb)

Helper function to get particle trajectories as a 3-dimensions array (N,M,T) instead of matrix of vectors.

LowLevelParticleFilters.update! — Function

ll, e = update!(f::AbstractFilter, u, y, p = parameters(f), t = index(f))

Perform one step of predict! and correct!, returns loglikelihood and prediction error

LowLevelParticleFilters.weigthed_cov — Method

weigthed_cov(x,we)

Similar to weigthed_mean, but returns covariances

LowLevelParticleFilters.weigthed_mean — Method

x̂ = weigthed_mean(x,we)

Calculated weighted mean of particle trajectories. we are expweights.

LowLevelParticleFilters.weigthed_mean — Method

x̂ = weigthed_mean(pf)
x̂ = weigthed_mean(s::PFstate)

LowLevelParticleFilters.rk4 — Function

f_discrete = rk4(f, Ts; supersample = 1)

Discretize a continuous-time dynamics function f using RK4 with sample time Ts. f is assumed to have the signature f : (x,u,p,t)->ẋ and the returned function f_discrete : (x,u,p,t)->x(t+1).

supersample determins the number of internal steps, 1 is often sufficient.