Exported functions and types

Index

LowLevelParticleFilters.AdvancedParticleFilter

LowLevelParticleFilters.KalmanFilter

LowLevelParticleFilters.ParticleFilter

LowLevelParticleFilters.SigmaFilter

LowLevelParticleFilters.TupleProduct

LowLevelParticleFilters.UnscentedKalmanFilter

LowLevelParticleFilters.commandplot

LowLevelParticleFilters.correct!

LowLevelParticleFilters.debugplot

LowLevelParticleFilters.densityplot

LowLevelParticleFilters.dimensiondensity

LowLevelParticleFilters.forward_trajectory

LowLevelParticleFilters.forward_trajectory

LowLevelParticleFilters.log_likelihood_fun

LowLevelParticleFilters.loglik

LowLevelParticleFilters.logsumexp!

LowLevelParticleFilters.mean_trajectory

LowLevelParticleFilters.metropolis

LowLevelParticleFilters.predict!

LowLevelParticleFilters.reset!

LowLevelParticleFilters.simulate

LowLevelParticleFilters.smooth

LowLevelParticleFilters.smooth

LowLevelParticleFilters.smoothed_cov

LowLevelParticleFilters.smoothed_mean

LowLevelParticleFilters.smoothed_trajs

LowLevelParticleFilters.trajectorydensity

LowLevelParticleFilters.update!

LowLevelParticleFilters.weigthed_cov

LowLevelParticleFilters.weigthed_mean

LowLevelParticleFilters.weigthed_mean

LowLevelParticleFilters.AdvancedParticleFilter — Method

AdvancedParticleFilter(Nparticles, dynamics, measurement, measurement_likelihood, dynamics_density, initial_density; kwargs...)

LowLevelParticleFilters.KalmanFilter — Type

KalmanFilter(A,B,C,D,R1,R2,d0=MvNormal(R1))

The matrices A,B,C,D define the dynamics

x' = Ax + Bu + w
y  = Cx + Du + e

where w ~ N(0, R1), e ~ N(0, R2) and x(0) ~ d0

LowLevelParticleFilters.ParticleFilter — Method

ParticleFilter(num_particles, dynamics, measurement, dynamics_density, measurement_density, initial_density)

LowLevelParticleFilters.SigmaFilter — Method

SigmaFilter(N,dynamics,measurement,measurement_likelihood,df,d0)

LowLevelParticleFilters.TupleProduct — Type

TupleProduct(v::NTuple{N,UnivariateDistribution})

Create a product distribution where the individual distributions are stored in a tuple. Supports mixed/hybrid Continuous and Discrete distributions

LowLevelParticleFilters.UnscentedKalmanFilter — Type

UnscentedKalmanFilter(dynamics,measurement,R1,R2,d0=MvNormal(Matrix(R1)))

LowLevelParticleFilters.commandplot — Method

commandplot(pf, u, y; kwargs...)

Produce a helpful plot. For customization options (kwargs...), see ?pplot. After each time step, a command from the user is requested.

q: quit

s n: step n steps

LowLevelParticleFilters.correct! — Function

 ll = correct!(f, u, y, t = index(f))

Update state/covariance/weights based on measurement y, returns loglikelihood.

LowLevelParticleFilters.debugplot — Method

debugplot(pf, u, y; runall=false, kwargs...)

Produce a helpful plot. For customization options (kwargs...), see ?pplot.

runall=false: if true, runs all time steps befor displaying (faster), if false, displays the plot after each time step.

The generated plot becomes quite heavy. Initially, try limiting your input to 100 time steps to verify that it doesn't crash.

LowLevelParticleFilters.densityplot — Function

densityplot(x,[w])

Plot (weighted) particles densities

LowLevelParticleFilters.dimensiondensity — Function

dimensiondensity(pf,x,we,y, dimension, nbinsy=30, xreal=nothing)

Same as trajectorydensity but only plots subplot dimension.

LowLevelParticleFilters.forward_trajectory — Method

x,w,we,ll = forward_trajectory(pf, u::AbstractVector, y::AbstractVector)

Run the particle filter for a sequence of inputs and measurements. Return particles, weights, expweights and loglikelihood

If MonteCarloMeasurements.jl is loaded, you may transform the output particles to Matrix{MonteCarloMeasurements.Particles} using Particles(x,we). Internally, the particles are then resampled such that they all have unit weight. This is conventient for making use of the plotting facilities of MonteCarloMeasurements.jl.

LowLevelParticleFilters.forward_trajectory — Method

x,xt,R,Rt,ll = forward_trajectory(kf::AbstractKalmanFilter, u::Vector, y::Vector)

Run a Kalman filter forward

Returns:

x: predictions

xt: filtered estimates

R: predicted covariance matrices

Rt: filter covariances

ll: loglik

LowLevelParticleFilters.log_likelihood_fun — Method

ll(θ) = log_likelihood_fun(filter_from_parameters(θ::Vector)::Function, priors::Vector{Distribution}, u, y)

returns function θ -> p(y|θ)p(θ)

LowLevelParticleFilters.loglik — Method

ll = loglik(filter,u,y)

Calculate loglikelihood for entire sequences u,y

LowLevelParticleFilters.logsumexp! — Function

ll = logsumexp!(w, we [, maxw])

Normalizes the weight vector w and returns the weighted log-likelihood

https://arxiv.org/pdf/1412.8695.pdf eq 3.8 for p(y) https://discourse.julialang.org/t/fast-logsumexp/22827/7?u=baggepinnen for stable logsumexp

LowLevelParticleFilters.mean_trajectory — Method

x,ll = mean_trajectory(pf, u::Vector{Vector}, y::Vector{Vector})

This Function resets the particle filter to the initial state distribution upon start

LowLevelParticleFilters.metropolis — Function

metropolis(ll::Function(θ), R::Int, θ₀::Vector, draw::Function(θ) = naive_sampler(θ₀))

Performs MCMC sampling using the marginal Metropolis (-Hastings) algorithm draw = θ -> θ' samples a new parameter vector given an old parameter vector. The distribution must be symmetric, e.g., a Gaussian. R is the number of iterations. See log_likelihood_fun

Example:

filter_from_parameters(θ) = ParticleFilter(N, dynamics, measurement, MvNormal(n,exp(θ[1])), MvNormal(p,exp(θ[2])), d0)
priors = [Normal(0,0.1),Normal(0,0.1)]
ll     = log_likelihood_fun(filter_from_parameters,priors,u,y,1)
θ₀ = log.([1.,1.]) # Initial point
draw = θ -> θ .+ rand(MvNormal(0.1ones(2))) # Function that proposes new parameters (has to be symmetric)
burnin = 200 # If using threaded call, provide number of burnin iterations
# @time theta, lls = metropolis(ll, 2000, θ₀, draw) # Run single threaded
# thetam = reduce(hcat, theta)'
@time thetalls = LowLevelParticleFilters.metropolis_threaded(burnin, ll, 5000, θ₀, draw) # run on all threads, will provide (2000-burnin)*nthreads() samples
histogram(exp.(thetalls[:,1:2]), layout=3)
plot!(thetalls[:,3], subplot=3) # if threaded call, log likelihoods are in the last column

LowLevelParticleFilters.predict! — Function

predict!(f, u, t = index(f))

Move filter state forward in time using dynamics equation and input vector u.

LowLevelParticleFilters.reset! — Method

Reset the filter to initial state and covariance/distribution

LowLevelParticleFilters.simulate — Method

x,u,y = simulate(f::AbstractFilter,T::Int,du::Distribution, [N])

Simulate dynamical system forward in time, returns state sequence, inputs and measurements du is a distribution of random inputs.

If MonteCarloMeasurements.jl is loaded, the argument N::Int can be supplied, in which case N simulations are done and the result is returned in the form of Vector{MonteCarloMeasurements.Particles}.

LowLevelParticleFilters.smooth — Method

xT,RT,ll = smooth(kf::KalmanFilter, u::Vector, y::Vector)

Returns smoothed estimates of state x and covariance R given all input output data u,y

LowLevelParticleFilters.smooth — Method

xb,ll = smooth(pf, M, u, y)

Perform particle smoothing using forward-filtering, backward simulation. Return smoothed particles and loglikelihood. See also smoothed_trajs, smoothed_mean, smoothed_cov

LowLevelParticleFilters.smoothed_cov — Method

smoothed_cov(xb)

Helper function to calculate the covariance of smoothed particle trajectories

LowLevelParticleFilters.smoothed_mean — Method

smoothed_mean(xb)

Helper function to calculate the mean of smoothed particle trajectories

LowLevelParticleFilters.smoothed_trajs — Method

smoothed_trajs(xb)

Helper function to get particle trajectories as a 3-dimensions array (N,M,T) instead of matrix of vectors.

LowLevelParticleFilters.trajectorydensity — Function

trajectorydensity(pf,x,we,y, nbinsy=30, xreal=nothing)

Plots particle densities along trajectory. See the readme for an example.

LowLevelParticleFilters.update! — Function

ll, e = update!(f::AbstractFilter, u, y, t = index(f))

Perform one step of predict! and correct!, returns loglikelihood and prediction error

LowLevelParticleFilters.weigthed_cov — Method

Similar to weigthed_mean, but returns covariances

LowLevelParticleFilters.weigthed_mean — Method

x̂ = weigthed_mean(x,we)

Calculated weighted mean of particle trajectories. we are expweights.

LowLevelParticleFilters.weigthed_mean — Method

x̂ = weigthed_mean(pf)
x̂ = weigthed_mean(s::PFstate)